dbms-backend/requirements.txt
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.9
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
